// src/modules/cermon/cermon-report/interface/report-weekly.interface.ts
export interface IReportWeekly {
  region_id: number;
  region_created_by: Date;
  region_updated_by: Date;
  region_deleted_by: Date;
  region_created_at: Date;
  region_updated_at: Date;
  region_deleted_at: Date;
  region_name: string;
  region_alt_name: string;
  region_location: string;
  region_parent_id: number;
  cermon_unique_id: string;
  cermon_id: number;
  cermon_name: string;
  cermon_time: string;
  cermon_day: string;
  cermon_region_id: number;
  cermon_segment: string;
  cermon_description: string;
  cermon_created_by: Date;
  cermon_updated_by: Date;
  cermon_deleted_by: Date;
  cermon_created_at: Date;
  cermon_updated_at: Date;
  cermon_deleted_at: Date;
  report_id: number;
  report_created_by: Date;
  report_updated_by: Date;
  report_deleted_by: Date;
  report_created_at: Date;
  report_updated_at: Date;
  report_deleted_at: Date;
  report_cermon_id: string;
  report_region_id: number;
  report_date: number;
  report_total_male: number;
  report_total_female: number;
  report_total_new_male: number;
  report_total_new_female: number;
}